Changed type names.

+++ src/lwe/lwe_to_lwe_keyswitch.cpp
@@ -20,17 +20,17 @@ LWEToLWEKeySwitchKey& LWEToLWEKeySwitchKey::operator=(const LWEToLWEKeySwitchKey
  
 void LWEToLWEKeySwitchKey::key_switching_key_gen(std::shared_ptr<LWESK> sk_origin, std::shared_ptr<LWEGadgetSK> sk_dest){  
     key_content = std::unique_ptr<std::unique_ptr<LWEGadgetCT>[]>(new std::unique_ptr<LWEGadgetCT>[origin->dim]); 
-    for(int i = 0; i < origin->dim; ++i){      
+    for(int32_t i = 0; i < origin->dim; ++i){      
         key_content[i] = std::unique_ptr<LWEGadgetCT>(sk_dest->gadget_encrypt(sk_origin->key[i]));  
     }   
 } 
  
-void LWEToLWEKeySwitchKey::set_key_switch_type(long base, int digits){ 
-    long bits_Q = Utils::power_times(destination->modulus, 2);
-    long bits_base = Utils::power_times(base, 2);
-    long bits_N = Utils::power_times(origin->dim, 2);  
-    long sum_lazy_bits = bits_Q + bits_base + digits + bits_N;
-    long sum_partial_lazy_bits = bits_Q + bits_base + digits + 1;
+void LWEToLWEKeySwitchKey::set_key_switch_type(uint64_t base, int32_t digits){ 
+    uint32_t bits_Q = Utils::power_times(destination->modulus, 2);
+    uint32_t bits_base = Utils::power_times(base, 2);
+    uint32_t bits_N = Utils::power_times(origin->dim, 2);  
+    uint32_t sum_lazy_bits = bits_Q + bits_base + digits + bits_N;
+    uint32_t sum_partial_lazy_bits = bits_Q + bits_base + digits + 1;
     if(sum_lazy_bits < 64){ 
         ks_type = lazy_key_switch;  
     }else if(sum_partial_lazy_bits < 64){ 
@@ -52,7 +52,7 @@ void LWEToLWEKeySwitchKey::lwe_to_lwe_key_switch(LWECT *lwe_ct_out, LWECT *lwe_c
 void LWEToLWEKeySwitchKey::lwe_to_lwe_key_switch_lazy(LWECT *lwe_ct_out, LWECT *lwe_ct_in){ 
     key_content[0]->gadget_mul_lazy(lwe_ct_out, lwe_ct_in->ct[1]); 
     LWECT temp_lwe_ct(destination);
-    for(int i=2; i < origin->dim+1; ++i){  
+    for(int32_t i=2; i < origin->dim+1; ++i){  
         key_content[i-1]->gadget_mul_lazy(&temp_lwe_ct, lwe_ct_in->ct[i]);  
         lwe_ct_out->add_lazy(lwe_ct_out, &temp_lwe_ct);
     }  
@@ -63,7 +63,7 @@ void LWEToLWEKeySwitchKey::lwe_to_lwe_key_switch_lazy(LWECT *lwe_ct_out, LWECT *
 void LWEToLWEKeySwitchKey::lwe_to_lwe_key_switch_partial_lazy(LWECT *lwe_ct_out, LWECT *lwe_ct_in){ 
     key_content[0]->gadget_mul_lazy(lwe_ct_out, lwe_ct_in->ct[1]); 
     LWECT temp_lwe_ct(destination);  
-    for(int i=2; i < origin->dim+1; ++i){   
+    for(int32_t i=2; i < origin->dim+1; ++i){   
         key_content[i-1]->gadget_mul_lazy(&temp_lwe_ct, lwe_ct_in->ct[i]);  
         lwe_ct_out->add(lwe_ct_out, &temp_lwe_ct);
     }  
@@ -75,7 +75,7 @@ void LWEToLWEKeySwitchKey::lwe_to_lwe_key_switch_partial_lazy(LWECT *lwe_ct_out,
 void LWEToLWEKeySwitchKey::lwe_to_lwe_key_switch_bussy(LWECT *lwe_ct_out, LWECT *lwe_ct_in){ 
     key_content[0]->gadget_mul(lwe_ct_out, lwe_ct_in->ct[1]); 
     LWECT temp_lwe_ct(destination);
-    for(int i=2; i < origin->dim+1; ++i){  
+    for(int32_t i=2; i < origin->dim+1; ++i){  
         key_content[i-1]->gadget_mul(&temp_lwe_ct, lwe_ct_in->ct[i]);  
         lwe_ct_out->add(lwe_ct_out, &temp_lwe_ct);
     }
